Implement page reading for ps2 cards
This is basically a port of the `Card_ReadPageData` function from
ps3mca-tool.
Verifying checksums is still missing.

The dump mechanism trusts that the reported size of the card is correct,
so my 64Mb card which reports itself as 8Mb is not dumped in its
entirety. The fix for this is probably to implement a probe similar to
`mcman_findfree2` from mcman in ps2sdk.