Explain the calculation for entries per cluster
In ps3mca-tool, the value for entries per cluster is hard-coded to be
256 while mymc calculates it by multiplying the size of a page with the
number of pages per cluster and dividing by four. The value of four is
not explained there but the explanation is probably that the file system is
FAT32 where each entry is 32 bits and thereby four 8-bit bytes long. RR
ps2mc-fs (http://www.csclub.uwaterloo.ca:11068/mymc/ps2mcfs.html)
specifies that each entry in the file allocation table is 32 bits long.

https://en.wikipedia.org/wiki/Design_of_the_FAT_file_system